Trait tor_netdir::params::FromInt32Saturating

pub trait FromInt32Saturating {
    // Required method
    fn from_saturating(val: i32) -> Self;
}

An object that can be constructed from an i32, with saturating semantics.

Required Methods

fn from_saturating(val: i32) -> Self

Construct an instance of this object from val.

If val is too low, treat it as the lowest value that would be valid. If val is too high, treat it as the highest value that would be valid.