Struct tor_netdir::Relay

pub struct Relay<'a> { /* private fields */ }

A view of a relay on the Tor network, suitable for building circuits.

Implementations

impl<'a> Relay<'a>

pub fn id(&self) -> &Ed25519Identity

Return the Ed25519 ID for this relay.

pub fn rsa_id(&self) -> &RsaIdentity

Return the RsaIdentity for this relay.

pub fn same_relay(&self, other: &Relay<'_>) -> bool

Return true if this relay and other seem to be the same relay.

(Two relays are the same if they have the same identity.)

pub fn supports_exit_port_ipv4(&self, port: u16) -> bool

Return true if this relay allows exiting to port on IPv4.

pub fn supports_exit_port_ipv6(&self, port: u16) -> bool

Return true if this relay allows exiting to port on IPv6.

pub fn is_dir_cache(&self) -> bool

Return true if this relay is suitable for use as a directory cache.

pub fn is_flagged_guard(&self) -> bool

Return true if this relay is marked as a potential Guard node.

pub fn in_same_subnet( &self, other: &Relay<'_>, subnet_config: &SubnetConfig ) -> bool

Return true if both relays are in the same subnet, as configured by subnet_config.

Two relays are considered to be in the same subnet if they have IPv4 addresses with the same subnets_family_v4-bit prefix, or if they have IPv6 addresses with the same subnets_family_v6-bit prefix.

pub fn in_same_family(&self, other: &Relay<'_>) -> bool

Return true if both relays are in the same family.

(Every relay is considered to be in the same family as itself.)

pub fn policies_allow_some_port(&self) -> bool

Return true if there are any ports for which this Relay can be used for exit traffic.

(Returns false if this relay doesn’t allow exit traffic, or if it has been flagged as a bad exit.)

pub fn ipv4_policy(&self) -> Arc<PortPolicy>

Return the IPv4 exit policy for this relay. If the relay has been marked BadExit, return an empty policy

pub fn ipv6_policy(&self) -> Arc<PortPolicy>

Return the IPv6 exit policy for this relay. If the relay has been marked BadExit, return an empty policy

pub fn ipv4_declared_policy(&self) -> &Arc<PortPolicy>

Return the IPv4 exit policy declared by this relay. Contrary to Relay::ipv4_policy, this does not verify if the relay is marked BadExit.

pub fn ipv6_declared_policy(&self) -> &Arc<PortPolicy>

Return the IPv6 exit policy declared by this relay. Contrary to Relay::ipv6_policy, this does not verify if the relay is marked BadExit.