Trait tor_netdir::MdReceiver

pub trait MdReceiver {
    // Required methods
    fn missing_microdescs(&self) -> Box<dyn Iterator<Item = &MdDigest> + '_>;
    fn add_microdesc(&mut self, md: Microdesc) -> bool;
    fn n_missing(&self) -> usize;
}

A partial or full network directory that we can download microdescriptors for.

Required Methods

fn missing_microdescs(&self) -> Box<dyn Iterator<Item = &MdDigest> + '_>

Return an iterator over the digests for all of the microdescriptors that this netdir is missing.

fn add_microdesc(&mut self, md: Microdesc) -> bool

Add a microdescriptor to this netdir, if it was wanted.

Return true if it was indeed wanted.

fn n_missing(&self) -> usize

Return the number of missing microdescriptors.